+++ server/src/main/java/com/ones/admin/config/OpenApiConfig.java
@@ -25,7 +25,7 @@ public class OpenApiConfig {
 
     private final String applicationVersion;
 
-    public OpenApiConfig(@Value("${ones.version:v0.0.30}") String applicationVersion) {
+    public OpenApiConfig(@Value("${ones.version:v0.0.31}") String applicationVersion) {
         this.applicationVersion = applicationVersion;
     }